requirements.txt
requests>=2.20.0
pycryptodome>=3.6.6
